requirements.txt
fastapi==0.63.0
uvicorn
fastapi-utils==0.2.1
PyYAML==6.0
